@@ -2,7 +2,11 @@
 	import type { PageData } from './$types';
 	import PostCard from '$lib/components/PostCard.svelte';
 
-	export let data: PageData;
+	interface Props {
+		data: PageData;
+	}
+
+	let { data }: Props = $props();
 </script>
 
 <svelte:head>
